Split up the test and add randomness.
