+++ logstash-core/src/test/java/org/logstash/plugin/PluginIntegrationTest.java
@@ -6,13 +6,11 @@ import org.logstash.Event;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedList;
-import java.util.function.Consumer;
 
 import static org.junit.Assert.assertEquals;
 
 public class PluginIntegrationTest {
-    private Collection<Collection<Event>> batches = new LinkedList<>();
-
+    private int eventCount = Util.random.nextInt(100);
     static Collection<Event> generateEvents(int count) {
         Collection<Event> events = new LinkedList<>();
         for (int i = 0; i < count; i++) {
@@ -24,32 +22,6 @@ public class PluginIntegrationTest {
         return Collections.unmodifiableCollection(events);
     }
 
-    @Test
-    public void testInput() {
-        Input i = new TestInput();
-        i.run((events) -> batches.add(events));
-        assertEquals(batches.size(), 1);
-        for (Collection<Event> events : batches) {
-            assertEquals(events.size(), 10);
-        }
-    }
-
-    private class TestInput implements Input {
-        @Override
-        public void run(Consumer<Collection<Event>> consumer) {
-            Collection<Event> events = new LinkedList<>();
-
-            for (int i = 0; i < 10; i++) {
-                Event e = new Event();
-                e.setField("message", "hello world");
-                e.setField("i", i);
-                events.add(e);
-            }
-
-            consumer.accept(Collections.unmodifiableCollection(events));
-        }
-    }
-
     private class TestFilter implements Processor {
         @Override
         public Collection<Event> process(Collection<Event> events) {
@@ -58,6 +30,23 @@ public class PluginIntegrationTest {
             }
             return null;
         }
+    }
+
+    private Processor processor = new TestFilter();
+
+    @Test
+    public void testFilter() {
+        Collection<Event> events = generateEvents(eventCount);
+        Collection<Event> newEvents = processor.process(events);
+
+        assertEquals(null, newEvents);
+
+        int i = 0;
+        for (Event event : events) {
+            assertEquals(event.getField("i"), i);
+            i++;
+        }
+
     }
 
 }
