+++ logstash-core/src/test/java/org/logstash/plugin/Util.java
@@ -1,0 +1,8 @@
+package org.logstash.plugin;
+
+import java.util.Random;
+
+public class Util {
+    public static final Random random = new Random();
+
+}
